pyproject.toml
[project]
name = "trend-mirror"
version = "0.1.0"
description = "TrendMirror Agentic Workflow"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.112.0",
    "uvicorn[standard]>=0.30.0",
    "chromadb==1.3.7",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
    "langgraph",
    "langchain-upstage",
    "langchain-chroma",
    "langchain-community",
    "langchain-core",
    "langsmith",
    "tiktoken",     # 노트북에서 사용됨
    "reportlab",    # 노트북에서 PDF 생성용
    "requests",
    "langchain>=1.2.3",
]
